+++ R/formula-re.R
@@ -19,6 +19,12 @@
 #' @param id Optional character string. All group-level terms across the model
 #'   with the same \code{id} will be modeled as correlated (if \code{cor} is
 #'   \code{TRUE}). See \code{\link{brmsformula}} for more details.
+#' @param cov An optional matrix which is proportional to the withon-group
+#'   covariance matrix of the group-level effects. All levels of the grouping
+#'   factor should appear as rownames of the corresponding matrix. This argument
+#'   can be used, among others, to model pedigrees and phylogenetic effects. See
+#'   \code{vignette("brms_phylogenetics")} for more details. By default, levels
+#'   of the same grouping factor are modeled as independent of each other.
 #' @param dist Name of the distribution of the group-level effects.
 #' Currently \code{"gaussian"} is the only option.
 #' 
@@ -40,7 +46,8 @@
 #' }
 #' 
 #' @export
-gr <- function(..., by = NULL, cor = TRUE, id = NA, dist = "gaussian") {
+gr <- function(..., by = NULL, cor = TRUE, id = NA,
+               cov = NULL, dist = "gaussian") {
   label <- deparse(match.call())
   groups <- as.character(as.list(substitute(list(...)))[-1])
   if (length(groups) > 1L) {
@@ -58,9 +65,18 @@ gr <- function(..., by = NULL, cor = TRUE, id = NA, dist = "gaussian") {
   } else {
     by <- ""
   }
+  cov <- substitute(cov)
+  if (!is.null(cov)) {
+    cov <- all.vars(cov)
+    if (length(cov) != 1L) {
+      stop2("Argument 'cov' must contain exactly one variable.")
+    }
+  } else {
+    cov <- ""
+  }
   dist <- match.arg(dist, c("gaussian", "student"))
   allvars <- str2formula(c(groups, by))
-  nlist(groups, allvars, label, by, cor, id, dist, type = "")
+  nlist(groups, allvars, label, by, cor, id, cov, dist, type = "")
 }
 
 #' Set up multi-membership grouping terms in \pkg{brms}
@@ -105,8 +121,8 @@ gr <- function(..., by = NULL, cor = TRUE, id = NA, dist = "gaussian") {
 #' }
 #'   
 #' @export
-mm <- function(..., weights = NULL, scale = TRUE,
-               cor = TRUE, id = NA, dist = "gaussian") {
+mm <- function(..., weights = NULL, scale = TRUE, cor = TRUE, 
+               id = NA, cov = NULL, dist = "gaussian") {
   label <- deparse(match.call())
   groups <- as.character(as.list(substitute(list(...)))[-1])
   if (length(groups) < 2) {
@@ -117,6 +133,15 @@ mm <- function(..., weights = NULL, scale = TRUE,
   }
   cor <- as_one_logical(cor)
   id <- as_one_character(id, allow_na = TRUE)
+  cov <- substitute(cov)
+  if (!is.null(cov)) {
+    cov <- all.vars(cov)
+    if (length(cov) != 1L) {
+      stop2("Argument 'cov' must contain exactly one variable.")
+    }
+  } else {
+    cov <- ""
+  }
   dist <- match.arg(dist, c("gaussian", "student"))
   scale <- as_one_logical(scale)
   weights <- substitute(weights)
@@ -128,8 +153,8 @@ mm <- function(..., weights = NULL, scale = TRUE,
     weightvars <- str2formula(weightvars)
   }
   nlist(
-    groups, weights, weightvars, allvars, 
-    label, by = "", cor, id, dist, type = "mm"
+    groups, weights, weightvars, allvars, label, 
+    by = "", cor, id, cov, dist, type = "mm"
   )
 }
 
@@ -551,6 +576,7 @@ tidy_ranef <- function(bterms, data, all = TRUE, old_levels = NULL) {
       cor = re$cor[[i]],
       type = re$type[[i]],
       by = re$gcall[[i]]$by,
+      cov = re$gcall[[i]]$cov,
       dist = re$gcall[[i]]$dist,
       stringsAsFactors = FALSE
     )
@@ -639,6 +665,8 @@ tidy_ranef <- function(bterms, data, all = TRUE, old_levels = NULL) {
       # for newdata numeration has to depend on the original levels
       attr(ranef, "levels") <- old_levels
     }
+    # incorporate deprecated 'cov_ranef' argument
+    ranef <- update_ranef_cov(ranef, bterms)
   }
   # ordering after IDs matches the order of the posterior samples 
   # if multiple IDs are used for the same grouping factor (#835)
@@ -761,4 +789,71 @@ get_rnames <- function(ranef, group = NULL, bylevels = NULL) {
     out <- outer(out, bylabels, paste, sep = ":")
   }
   out
+}
+
+# validate within-group covariance matrices
+# @param M a matrix to be validated
+validate_recov_matrix <- function(M) {
+  M <- as.matrix(M)
+  if (!isSymmetric(unname(M))) {
+    stop2("Within-group covariance matrices must be symmetric.")
+  }
+  found_levels <- rownames(M)
+  if (is.null(found_levels)) {
+    found_levels <- colnames(M)
+  }
+  if (is.null(found_levels)) {
+    stop2("Row or column names are required for within-group covariance matrices.")
+  }
+  rownames(M) <- colnames(M) <- found_levels
+  evs <- eigen(M, symmetric = TRUE, only.values = TRUE)$values
+  if (min(evs) <= 0) {
+    stop2("Within-group covariance matrices must be positive definite.")
+  }
+  M
+}
+
+# check validity of the 'cov_ranef' argument
+# argument 'cov_ranef' is deprecated as of version 2.12.5
+validate_cov_ranef <- function(cov_ranef) {
+  is_valid <- isTRUE(attr(cov_ranef, "valid"))
+  if (is.null(cov_ranef) || is_valid) {
+    return(cov_ranef)
+  }
+  warning2(
+    "Argument 'cov_ranef' is deprecated and will be removed in the future. ",
+    "Please use argument 'cov' in function 'gr' instead."
+  )
+  cr_names <- names(cov_ranef)
+  cr_is_named <- length(cr_names) && all(nzchar(cr_names))
+  if (!is.list(cov_ranef) || !cr_is_named) {
+    stop2("'cov_ranef' must be a named list.")
+  }
+  if (any(duplicated(cr_names))) {
+    stop2("Names of 'cov_ranef' must be unique.")
+  }
+  structure(cov_ranef, valid = TRUE)
+}
+
+# update 'ranef' according to information in 'cov_ranef'
+# argument 'cov_ranef' is deprecated as of version 2.12.5
+update_ranef_cov <- function(ranef, bterms) {
+  cr_names <- names(bterms$cov_ranef)
+  if (!length(cr_names)) {
+    return(ranef)
+  }
+  unused_names <- setdiff(cr_names, ranef$group)
+  if (length(unused_names)) {
+    stop2("The following elements of 'cov_ranef' are unused: ",
+          collapse_comma(unused_names))
+  }
+  has_cov <- ranef$group %in% cr_names
+  ranef$cov[has_cov] <- ranef$group[has_cov]
+  ranef
+}
+
+# extract 'cov_ranef' for storage in 'data2'
+# @param x a list-like object
+get_data2_cov_ranef <- function(x) {
+  x[["cov_ranef"]]
 }